% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grpslope.R
\name{grpSLOPESolver}
\alias{grpSLOPESolver}
\title{Group SLOPE algorithm}
\usage{
grpSLOPESolver(X, Dinv, b, group, lambda, opts = list())
}
\description{
Compute the coefficient estimates of the Group SLOPE model
}

