% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grpslope.R
\name{proximalGradientSolverGroupSLOPE}
\alias{proximalGradientSolverGroupSLOPE}
\title{Proximal gradient method for Group SLOPE}
\usage{
proximalGradientSolverGroupSLOPE(y, A, group, wt, lambda, max.iter = 10000,
  verbose = TRUE, tolerance = 1e-06, x.init = vector())
}
\arguments{
\item{y}{the response vector}

\item{A}{the model matrix}

\item{group}{A vector describing the grouping structure. It should 
contain a group id for each predictor variable.}

\item{wt}{A vector of weights}

\item{lambda}{A decreasing sequence of regularization parameters \eqn{\lambda}}

\item{max.iter}{Maximal number of iterations to carry out}

\item{verbose}{A \code{logical} specifying whether to print output or not}

\item{tolerance}{The tolerance used in the stopping criteria}

\item{x.init}{An optional initial value for the iterative algorithm}
}
\value{
A list with members:
  \describe{
    \item{x}{Solution (n-by-1 matrix)}
    \item{status}{Convergence status: 1 if optimal, 2 if iteration limit reached}
    \item{L}{Approximation of the Lipschitz constant (step size)}
    \item{iter}{Iterations of the proximal gradient method}
    \item{L.iter}{Total number of iterations spent in Lischitz search}
  }
}
\description{
Compute the coefficient estimates for the Group SLOPE problem.
}
\details{
\code{proximalGradientSolverGroupSLOPE} computes the coefficient estimates
for the Group SLOPE model. The employed optimization algorithm is FISTA with 
backtracking Lipschitz search.
}

