% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getGroupID}
\alias{getGroupID}
\title{Get a groupID object}
\usage{
getGroupID(group)
}
\arguments{
\item{group}{A vector describing the grouping structure. It should 
contain a group id for each predictor variable.}
}
\value{
An object of class groupID, which is a list, whose members are 
   vectors of indices corresponding to each group. The names of
   the list members are the corresponding group names.
}
\description{
Intended for internal use only.
}

