% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grpslope.R
\name{grpSLOPE}
\alias{grpSLOPE}
\title{Group SLOPE (Group Sorted L-One Penalized Estimation)}
\usage{
grpSLOPE(X, y, group, fdr, lambda, sigma = NULL,
  n.MC = floor(length(unique(group))/2), MC.reps = 5000, verbose = FALSE)
}
\arguments{
\item{X}{The model matrix}

\item{y}{The response variable}

\item{group}{A vector describing the grouping structure. It should 
contain a group id for each predictor variable.}

\item{fdr}{Target false discovery rate}

\item{lambda}{Method used to obtain the regularizing sequence lambda. Possible
values are "BH", "gaussian", "gaussianMC", "chiOrthoMax", "chiOrthoMean",
"chiEqual", "chiMean", "chiMC". See \code{\link{lambdaGroupSLOPE}} for detail.}

\item{sigma}{Noise level. If ommited, estimated from the data. See details.}

\item{n.MC}{When \code{method} is "gaussianMC" or "chiMC", the corrections of the entries of lambda will be 
computed up to the index given by \code{n.MC} only. See details.}

\item{MC.reps}{The number of repetitions of the Monte Carlo procedure}

\item{verbose}{Verbosity}
}
\value{
A list with members:
  \describe{
    \item{beta}{Solution vector}
    \item{selected}{Names of selected groups (i.e., groups of predictors with at least one coefficient estimate >0)}
    \item{optimal}{Convergence status}
    \item{iter}{Iterations of the proximal gradient method}
    \item{lambda}{Regularizing sequence}
    \item{lambda.method}{Method used to construct the regularizing sequence}
    \item{sigma}{(Sequence of) (estimated) noise levels used}
  }
}
\description{
Performs selection of significant groups of predictors and estimation of the
corresonding coefficients using the Group SLOPE method (see Brzyski et. al. (2015)
and Gossmann et. al. (2015)).
}
\details{
Multiple methods are available to generate the regularizing sequence \code{lambda},
see \code{\link{lambdaGroupSLOPE}} for detail.
If \code{method} is one of "chiOrthoMax", "chiOrthoMean",  "chiEqual", "chiMean", "chiMC",
then the model matrix is transformed by orthogonalization within each group (see Section 2.1
in Brzyski et. al.), and penalization is imposed on \eqn{\| X_{I_i} \beta_{I_i} \|}.
For other methods penalization is imposed directly on \eqn{\| \beta_{I_i} \|},
as in Gossmann et. al. (2015).
When \code{method} is "gaussianMC" or "chiMC", the corrections of the entries of lambda will be 
computed up to the index given by \code{n.MC} only. \code{n.MC} should be
less than or equal to \code{n.group}. Since lambda sequences obtained via MC tend to
flatten out quickly, it is reasonable to choose \code{n.MC} to be much smaller than the
number of groups.
}
\references{
A. Gossmann, S. Cao, Y.-P. Wang (2015), \emph{Identification of Significant Genetic Variants via SLOPE, and Its Extension to Group SLOPE}, \url{http://dx.doi.org/10.1145/2808719.2808743}

D. Brzyski, W. Su, M. Bogdan (2015), \emph{Group SLOPE — adaptive selection of groups of predictors}, \url{http://arxiv.org/abs/1511.09078}
}

