% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grpslope.R
\name{lambdaGroupSLOPE}
\alias{lambdaGroupSLOPE}
\title{Regularizing sequence for Group SLOPE}
\usage{
lambdaGroupSLOPE(q = 0.1, n.group = NULL, group = NULL, A = NULL,
  method = "gaussian", n.MC = n.group, MC.reps = 5000)
}
\arguments{
\item{q}{The nominal level for the false discovery rate}

\item{group}{A vector describing the grouping structure. It should 
contain a group id for each predictor variable.}

\item{A}{The model matrix}

\item{method}{Possible values are "gaussian", "chi", "gaussianMC", "chiMC"}

\item{n.MC}{The corrections of the entries of lambda will be 
computed up to the index given by \code{n.MC} only.}

\item{MC.reps}{The number of repetitions of the Monte Carlo procedure}

\item{n.groups}{Number of groups}
}
\description{
Generate the regularizing sequence \code{lambda} for the Group SLOPE
problem according to one of multiple methods.
}

