% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grpslope.R
\name{proxGroupSortedL1}
\alias{proxGroupSortedL1}
\title{Prox for group SLOPE}
\usage{
proxGroupSortedL1(y, group, lambda, method)
}
\arguments{
\item{y}{The response vector}

\item{group}{A vector describing the grouping structure. It should contain a group id
for each predictor variable.}

\item{lambda}{A decreasing sequence of regularization parameters \eqn{\lambda}.}

\item{method}{Specifies which implementation of the Sorted L1 norm prox should be used. 
See \code{\link[SLOPE]{prox_sorted_L1}}.}
}
\description{
Evaluate the proximal mapping for the group SLOPE problem.
}
\details{
\code{proxGroupSortedL1} evaluates the proximal mapping of the group SLOPE problem
by reducing it to the prox for the (regular) SLOPE and then applying the function
\code{\link[SLOPE]{prox_sorted_L1}}.
}
\examples{
grp <- c(0,0,0,1,1,0,2,1,0,2)
proxGroupSortedL1(y = 1:10, group = grp, lambda = 10:1)

}

