% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grpslope.R
\name{proximalGradientSolverGroupSLOPE}
\alias{proximalGradientSolverGroupSLOPE}
\title{Proximal gradient method for Group SLOPE}
\usage{
proximalGradientSolverGroupSLOPE(y, A, group, wt, lambda, max.iter = 10000,
  verbose = TRUE, tolerance = 1e-06, x.init = vector(), method = "rcpp")
}
\arguments{
\item{y}{the response vector}

\item{A}{the model matrix}

\item{group}{A vector describing the grouping structure. It should 
contain a group id for each predictor variable.}

\item{wt}{A vector of weights}

\item{lambda}{A decreasing sequence of regularization parameters \eqn{\lambda}}

\item{max.iter}{Maximal number of iterations to carry out}

\item{verbose}{A \code{logical} specifying whether to print output or not}

\item{tolerance}{The tolerance used in the stopping criteria}

\item{x.init}{An optional initial value for the iterative algorithm}

\item{method}{Specifies which implementation of the Sorted L1 norm prox should be used. 
See \code{\link{proxGroupSortedL1}} for detail.}
}
\value{
A list with members:
  \describe{
    \item{x}{Solution (n-by-1 matrix)}
    \item{status}{Convergence status: 1 if optimal, 2 if iteration limit reached}
    \item{L}{Approximation of the Lipschitz constant (step size)}
    \item{iter}{Iterations of the proximal gradient method}
    \item{L.iter}{Total number of iterations spent in Lischitz search}
  }
}
\description{
Compute the coefficient estimates for the Group SLOPE problem.
}
\details{
\code{proximalGradientSolverGroupSLOPE} computes the coefficient estimates
for the Group SLOPE model. The employed optimization algorithm is FISTA with 
backtracking Lipschitz search.
}
\examples{
set.seed(1)
A   <- matrix(runif(100, 0, 1), 10, 10)
grp <- c(0, 0, 1, 1, 2, 2, 2, 2, 2, 3)
wt  <- c(0.5, 0.5, 0.5, 0.5, 0.2, 0.2, 0.2, 0.2, 0.2, 1)
x   <- c(0, 0, 5, 1, 0, 0, 0, 1, 0, 3)
y   <- A \%*\% x
lam <- 0.1 * (10:1)
result <- proximalGradientSolverGroupSLOPE(y=y, A=A, group=grp, wt=wt, lambda=lam, 
                                           tolerance=1e-12, verbose=FALSE)
result$x

}
\references{
M. Bogdan, E. van den Berg, C. Sabatti, W. Su, E. Candes (2015), \emph{SLOPE - Adaptive variable selection via convex optimization}, \url{http://arxiv.org/abs/1407.3824}
}

