% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{proxSortedL1Rcpp}
\alias{proxSortedL1Rcpp}
\title{Fast prox for the Sorted L1 norm}
\usage{
proxSortedL1Rcpp(y, lambda)
}
\description{
A stack-based algorithm for the prox for the Sorted L1 norm,
which solves the problem in O(n) flops.
}

